Concluída tarefa: Importar ficheiro JSON Verificações feitas no ficheiro JSON para não enviar para BD special chars.

+++ HotelAgencyPT/src/main/java/hotel/agencypt/Controller/F_GerirStock.java
@@ -83,7 +83,7 @@ public class F_GerirStock implements Initializable {
     public void initialize(URL location, ResourceBundle resources) {
         String pid;
         String pd;
-        int qtd;
+        float qtd;
         String tipo_qtd;
         float preco;
         float vat;
